[guiconnect] - provide additional effect control params via DTO
This is a preparation for the next step, attempting to get rid of
extraneous UI state related to the control parameters (effect number,
effect type, routing bypass). Rather, these shall be provided alongside
with the effect parameters in the Data Transfer Object.

Some casual clean-up:
- "short int" is an exotic data type nowadays and requires additional
  assembly to be emitted for arithmetic and to access a "packed" array.
  All existing usages convert to int or compare to int constants.
- the UI for the Part-effect routing has special case treatment for uchar(127),
  which is then mapped to the default setting (1 == route to part out).
  I can not see any justification for that in the existing usages. In Synth,
  this value is only ever checked against the fixed constants 0 ("Next Effect")
  or 2 ("Dry Out"). When reading from XML, the value is conformed 0 <= x < 3

+++ src/Misc/SynthEngine.h
@@ -237,7 +237,7 @@ class SynthEngine
         EffectMgr *insefx[NUM_INS_EFX]; // insertion
 
         // part that's apply the insertion effect; -1 to disable
-        short int Pinsparts[NUM_INS_EFX];
+        int Pinsparts[NUM_INS_EFX];
 
         // connection to the currently active effect UI
         GuiDataExchange::Connection<EffectDTO> sysEffectUiCon;
